apollo_propeller: extract behaviour test helpers into test util module
## Goal
Keep behaviour_test.rs focused on tests by moving the shared TestNode / TestEnvironment
harness into a dedicated test util module (review feedback: the test file should contain
only tests).

## Summary of changes
- Add behaviour_test_utils.rs holding TestNode and TestEnvironment and their helpers,
  exposed as pub(crate) with a module-level #![allow(dead_code)] since not every test uses
  every helper.
- behaviour_test.rs now imports TestEnvironment from the util module and contains only the
  test cases.
- Wire mod behaviour_test_utils into lib.rs under #[cfg(test)].

## Key decision points
- Placed the extraction at the top of the propeller test stack rather than amending the
  infrastructure PR: a file move in the base PR would conflict with the three test commits
  that modify behaviour_test.rs on top of it.
- Used #![allow(dead_code)] at the module level instead of per-item #[expect(dead_code)]:
  the expect form errors once an item becomes used by a test, which is fragile across the
  stack as tests are added.
- Kept purely-internal helpers private; only the helpers called from the test module are
  pub(crate).
