+++ crates/apollo_propeller/src/lib.rs
@@ -21,6 +21,8 @@ pub mod unit_validator;
 #[cfg(test)]
 mod behaviour_test;
 #[cfg(test)]
+mod behaviour_test_utils;
+#[cfg(test)]
 mod merkle_test;
 #[cfg(test)]
 mod padding_test;
